requirements.txt
python-dotenv
gradio
gradio-multimodalchatbot
langchain-core
langchain
Pillow
langchain-community
transformers
torch
accelerate
requests
